Revert "Format our UTC dates in the user's current timezone"
This reverts 9a808c5a5607c27d182cfdb155c63f78addb9f04 [formerly d4ff6f30daebd6168ee83b8ea085880a90cbf388] as
that commit was incorrect.  We actually had a timezone
translation occuring in gwtjsonrpc that we didn't want,
so the Timestamp objects were already in the user's own
timezone, which is why we had to use the reverse version
of the offset to get them to display right.

Now that gwtjsonrpc is fixed, we shouldn't be doing this
at display time.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b43670882d6575312478d3e7dd00463609167eab