Define the special project 0 so admins can edit its rights
The special wildcard project 0 is now defined as '-- All Projects --'
so administrators can edit the project rights.  This of course means
that a project cannot be called '-- All Projects --' and that the
Git repository '-- All Projects --.git' is also off limits, but I
think both are reasonable limitations given how silly the name is.

Over SSH the wildcard project cannot be accessed, even though it
exists in the database.

The wildcard project's description and rights can be modified, but
the ownership of the wildcard project cannot be given away to another
group.  This prevents administrators from making the mistake of giving
out ownership of the wildcard project and having another user grant
very permissive rights to the entire world.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: d37f7106a37433ded4ad0b5818207387a40ad0c4