Support per-project access controls by group
Projects (and their changes) may now be restricted to being visible
only to certain groups, providing for shared but private access when
necessary.  These restrictions are enforced on the web and on the SSH
daemon on a per-project basis.

By default read access is granted to Anonymous Users, making the site
visible to all, but this can be overridden within a specific project,
or globally by removing the Anonymous Users READ grant from the default
project that everyone inherits from.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a49c3c6ecefb809ed461383c32e40ffd9df91889