Support other OpenID providers beyond Google Accounts
The Gerrit login form now displays an OpenID login page, allowing
the user to enter the URL of their OpenID provider, so Gerrit can
communicate with that server for login support.

Quick links to two popular login services, Google Accounts and
Yahoo! ID are supplied, to make it easier for users to connect
as one of these identities.

Yahoo! posed a problem as it does not permit appearing inside
of an <iframe> tag, especially on Safari or Firefox 3.  We now
only keep the SSO provider in an <iframe> if we know they are
going to be able to use the iframe, otherwise we toss them to
the toplevel window and save the current history token so the
page can be reloaded when we come back.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 8a129c4fbb3b3b30b36733e26d20a666600eea6c