Refactor the entire login panel to avoid reloading the module
Its significantly faster if the open id login panel is part of the
same application, and simply inserts itself into the sign in dialog
rather than loading through a separate HTML host page.

We also now remember (at the user's option) the last OpenID URL the
user successfully logged in with, so it can be presented in future
sign-in attempts.  This helps in case the user forgets which provider
they have used, or which account at that provider.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 9a2a97d6aa5fb9dfaed1a0870e0bdaa38ff803a5