Clean up the Gerrit1 account migration to be more limited
We only need to import accounts from Gerrit1 for a limited time, and
only on servers which have had the migration take place.  Most new
Gerrit2 installs won't need this function, so we don't want to drag
down the new account creation code with the extra query.

The allow_google_account_upgrade flag in system_config controls the
second query during account creation.  We also now drop the "?" from
the external_id strings when importing these accounts, as it doesn't
give us any additional value over not having it.  Finally we only let
the upgrade succeed if there is only one account with the given email,
if there are multiple accounts we abort and consider it to be new.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: f2fbd4889ea53c5cab7f8cb1a205f9fd6efc7ecf