Disable any use of cross site requests at the current time
We're starting to add a lot of cases where the output of a method
is dependent upon what the current user can see, like per-project
access controls.  Until we have that code protected by decision
logic that requires XSRF when the results differ for the current
user and the anonymous user we must ensure the XSRF token is used
to make the request.

Its far simpler at this point in time to just require XSRF keys
throughout the call interface, and make the browser run a few
more requests (to get or refresh the token, then run the method).

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 82239b7dfe924e63a811ea28d55a03da9cd7693f