Fix the unified patch display to avoid showing duplicate comments
The backend service for the unified patch formatter combined a
comment into the output multiple times, rather than only for the
first time that side was shown in the diff.  By removing it from
the cache we only show it the first time the line occurs, which
is where the user would expect it to be.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2711645183752c57b530108ce02182bf249f842a