Hide/show active comment editors upon sign out/sign in
They can't save the editor's contents without their cookie, so we
need to throw the editor away.

Upon sign in again we need to make sure the current set of draft
editors is valid.  We load all current drafts back in, allowing
the user to resume editing their unpublished comments.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 57fd9e38566a0e153f9bd911641066cb81ec6f5c