Allow users to register additional email addresses
The SSH username cannot be set without having an email address,
and some types of CLAs may require an email address to be set
in the contact information for the user.

However, not all OpenID providers track or transmit emails to
relying parties, so users must still be able to register an
address externally from their OpenID system.  We now send out
a validation email and use a secure token to verify the address
matches where the user sent it to.  A "mailto:" URL is used in
the external id table to register the additional address, thus
making it available as a choice for the user in the settings.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: e76e0ab31b03a65fab8435ec170fe55ced93f395