Add lists of all open, merged, abandoned changes on the server
To make pagination relatively efficient on the server side we use
a new "sort_key" column in the changes table.  This is a hex string
comprised of the pair lastUpdatedOn,changeId, creating a unique key
we can position in the index on.  Queries operate by performing a
range scan on the index, making them quite simple to execute.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 3cd1a9e395c34acb5fc6dea2793a363b5ced7657