Add a schema version table to facilitate live upgrades
The current version of the database schema is stored into the
schema_version.version_nbr field.  Future versions of Gerrit2
can update this to a different value and detect at startup if
an upgrade is required.  Upgrades would involve adding columns
to tables, or possibly manipulating data to conform to a new
convention.  The actual code required to perform an upgrade
isn't here yet, but we abort if the schema version isn't
recognized by the server.  This prevents an older Gerrit2 from
using a newer schema (once a newer schema is developed).

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: f5a97c4ce663accd87c6a260be5a2855c81b6452