Organize hot tables by their primary key when possible
These tables typically are accessed only by range scans on the
primary key, so storing rows in the same order they appear in
the index can benefit the server by ensuring the next row will
already be in cache when the index wants to read it.

Some database servers (e.g. Oracle, Informix, Sybase, SQL Server)
have a concept of "index organized table", where the data tuples
are directly embedded in the primary key index.  This is basically
what we want, but PostgreSQL doesn't offer that exact feature.  The
best we can do is tell the server of our desire to keep these rows
together, and run "CLUSTER" every so often so it can rebuild the
database over again.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 10a9f0f73be3e009d688b70725399fe024426655