Upgrade OpenID provider dyuproject to 1.1.1
This version contains built-in support for Google Account discovery,
so we don't need to hack it ourselves anymore.

Unfortunately the authentication approach used by RelyingParty assumes
you want to do an HTTP redirect to start the authentication.  Instead
Gerrit2 wants to send JSON to the client with the redirect URL.  We
must therefore code a little bit more on our side to ensure the user
manager has saved its cookie.

The library now also includes a utility map to create a URL encoded
string; we use this rather than StringBuilder append calls to build
up any URL which contains query parameters.

I'm specifically not using the SReg support in the library as it does
a request for additional properties we aren't too interested in, like
the user's gender.  Instead we just keep our own implementation of the
SReg extension.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: e0888d9c1932575dcfa7c8986da7278c24057aa6