Add an optional Account.Id to UserIdentity within PatchSetInfo
This way we can link to a user's dashboard if they were the one who
authored or committed a change.  Due to the way Git commits are made
not every email address field will necessarily match up to a Gerrit
account identity, so the field is optional and the client UI doesn't
draw a hyperlink of the account field isn't populated.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 42a57ba007992940674a3f3db7934701e04b7181