Always address comments to the patch set uploader/author/committer
This way comments get to people who are responsible for the change,
as typically the comments are about the code itself, so it should
go back to those who introduced it into the system for review.

We only send email to the commit author and/or committer if they
already had a Gerrit user account registered at the time that the
patch set was uploaded.  This tries to void spamming 3rd parties
who are outside of the review system, yet whose identity was used
in the author field of a commit (e.g. a patch picked up off some
mailing list that is suitable for inclusion).

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 50bcd05bf2953763f807c0425ec2cc61035189a3