Allow users to configure per-project email settings
The settings are tied to the "project watch", as this is a great
place to record what the user's project specific preferences are,
especially when it comes to notifications.

The settings don't actually do anything yet, but the UI is in
place and updates are written back to the database.

On import from Gerrit1 project owners and reviewers with a
Code Review status of 2 (aka max approval status) have any
imported project watch set to notifyNewChanges, as we can
reasonably guess they want to receive notifications when a
new change is uploaded for projects they have previously
said they are interested in.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: bf58b30bc072f27779780d8fab15165c2c04031f