Fix the "bouncing line" UI problem in patch views when using the cursor
When using the cursor to move up and down lines in the patch views
the rows visually jumped to be taller/shorter as the cursor moved
through them.  The problem was two issues:  the icon cell height
and the borders.

Cells without borders were 2px shorter then the active cells with
a border, so the text jumped around in the cell when the border is
added or removed.  Applying a transparent border to all cells on
the top and bottom fixes the problem, as the border size then does
not depend on the active status of the row.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 13fe99a2fc0dd1ec300116ae7fbe16d7cc068979