Don't include the footer "--" line in emails if the footer is empty
We should only include the "--\n" line in an email sent from Gerrit
if there will be content below that line.  Otherwise its just noise
and looks out of place in the message.

Usually there is content below the line, as the canonical_url is set
and thus we can make backlinks to ourselves.  But this isn't always
the case, so we should format a bit better when it isn't.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 5ff9d57feb6420370cc646dc3a3934f3ef4f3ad4