Add new project rights to allow direct branch updates over SSH
Four new rights can be used to control access over the SSH server:

 * Push Annotated Tag (refs/tags/*)
 * Push Branch:       (refs/heads/*)
   - Update Existing
   - Create New
   - Force Update (aka non-fast forward) ; Delete

These can be granted on a per-group/per-project basis to permit users
to remotely control branches over the Git SSH interface.  Its mostly
useful for groups that are used to having a standard Git server, where
"anything goes" is permitted within the scope of the project.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 20f6e9bdfd81cbd92ed04d7628bc1688da17f43e