Fix the subject column in change tables to truncate when too long
Because the text is actually inside a gwt Hyperlink object its down
inside of a div, not the td directly.  We need to apply the font
controls onto the div, otherwise the div overrides them and we don't
get the change subject to truncate nicely when its too long for the
available screen space.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 3d61ff743a9fdccfeeb09c73a204c1f548e53c4a