Clip all change subject lines at 80 columns in change tables
Firefox 3.0.6 and Safari 3.2.1 don't honor the overflow:hidden
style we use for the subject column of the change table.  All
other browsers (Opera, Firefox 2, Camino) honor this style and
clip the subject line rather than making the table wider than
the browser window.  Firefox 3 and Safari however are making the
table as wide as necessary to display everything, which results
in the trailing columns being place off screen when someone uses
a subject that is longer than expected.

Until I find a workaround in CSS that allows all browsers to do
automatic resizing of this column, I'm artificially clipping it
at 80 characters during display.  This is wide enough that any
reasonable message will fit without clipping, but not so wide as
to cause the tables to be unusable on a laptop display.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 719986649a505650a12133a0cca26456c1ae91db