Make the magic refs/heads/ constant available in GWT
JGit defines Constants.R_HEADS to be "refs/heads/" but we don't
have JGit's source code in our build path, so GWT can't use that
constant here in Branch.  It also can't process the Constants in
JGit as that class contains references to types that GWT lacks,
even though that would all strip out anyway during the build (as
it is never called).  So we duplicate these simple definitions.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2fe632e92c4cc87a7eedb229be288c347c5f165a