Add additional reviewers to an existing change
Additional reviewers can be requested to look at a change by entering
their name/email address into an input box below the approvals
table on a change screen. Any added reviewers are sent an email
with a link back to the change, so they can come visit Gerrit and
do the review. A score of 0 is also automatically assigned so the
user appears in the approval table, and the change shows in their
dashboard.

Bug: GERRIT-37
Signed-off-by: Brad Larson <bklarson@gmail.com>

Former-commit-id: bbc91c7dd6cd6b3e5ab1a123c7773e7bd1455005