Include "a=commit" in direct gitweb commit links
By including the commit action in the gitweb link for a commit the
gitweb CGI has a better idea of what we want it to show for this
object.  If the object isn't present yet (e.g. due to a mirroring
lag) we get a little better error message shown to the client.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: b407f7b94cd3fb7e417808dd8d454830db969b02