Don't bother logging IO errors caused by disappearing clients
If the client has broken off the network connection and that is why
we are throwing an exception up at the caller, its not really an
exception to be logged for the administrator.  Its a fairly harmless
case where the client started the network transfer, then just died.

We should consider sending a patch upstream to MINA SSHD to try and
use a more specific exception type here than just IOException, e.g.
EOFException, or modify read to return -1.  This might improve how
we trap the exception, and eventually ignore it.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2335b6dd60b6b034344667bccfbb4a2df46495e6