Allow users to delete OpenID identities no longer used
A user may desire to delete an OpenID identity they had previously
stored into their account, such as if they no longer trust that
provider, or are no longer affiliated with that provider.

To prevent the user from locking themselves out of their own user
account in Gerrit we forbid deleting the identity they last used
to login to the site under.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: d3946b140d0bd3fc2ca0625f35b452ac7a2235b4